renaming things, slim down endpoint url
