+++ domain-cc/cc-api-controller/build.gradle
@@ -18,7 +18,6 @@ dependencies {
   // compileOnly group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
 
   // Used by Controller classes
-  // implementation project(':domain-redo:redo-shared')
   // api project(':shared:lib-camel-connector')
   implementation project(':shared:lib-camel-connector')
