+++ domain-cc/cc-api-controller/src/main/java/gov/va/vro/api/cc/ResourceException.java
@@ -1,4 +1,4 @@
-package gov.va.vro.api.redo;
+package gov.va.vro.api.cc;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
