+++ domain-cc/cc-api-controller/src/main/java/gov/va/vro/api/cc/v3/CCResource.java
@@ -1,6 +1,6 @@
-package gov.va.vro.api.redo.v3;
+package gov.va.vro.api.cc.v3;
 
-import gov.va.vro.api.redo.ResourceException;
+import gov.va.vro.api.cc.ResourceException;
 import io.micrometer.core.annotation.Timed;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import javax.validation.Valid;
 
 @RequestMapping(value = "/v3", produces = "application/json")
-@Timed("redo") // See https://www.baeldung.com/micrometer
-@Tag(name = "Redo Domain")
-public interface RedoResource {
-  @PostMapping("/domain-redo/cmd/{endpoint}")
+@Timed("cc") // See https://www.baeldung.com/micrometer
+@Tag(name = "CC Domain")
+public interface CCResource {
+  @PostMapping("/domain-cc/{endpoint}")
   @Timed(value = "resource.post")
   ResponseEntity<ResourceResponse> callEndpoint(
           @PathVariable(value = "endpoint") String endpoint,
