+++ settings.gradle
@@ -98,7 +98,7 @@ include ':mocks:mock-bip-claims-api'
 include ':mocks:mock-mas-api'
 include ':mocks:mock-lighthouse-api'
 include ':mocks:mock-slack'
-include ':domain-cc:redo-api-controller'
+include ':domain-cc:cc-api-controller'
 include ':domain-xample:xample-api-controller'
 include ':domain-xample:xample-shared'
 include ':domain-xample:xample-workflows'
